requirements.txt
pipecatcloud
-e pipecat-march-openai[cartesia,openai,silero,deepgram,daily]
fastapi
uvicorn
python-dotenv
aiohttp
